+++ src/state/application-alerting/sagas.js
@@ -2,5 +2,9 @@ import actionCreators from './action-creators';
 import { api } from '../../api/application-alerting';
 import { alertingSaga } from '../alerting-utils/sagas';
 
-const componentSaga = alertingSaga('APPLICATION_ALERTING', actionCreators, api);
-export default componentSaga;
+const applicationAlertingSaga = alertingSaga(
+  'APPLICATION_ALERTING',
+  actionCreators,
+  api
+);
+export default applicationAlertingSaga;
