+++ src/state/environment-alerting/sagas.js
@@ -2,5 +2,9 @@ import actionCreators from './action-creators';
 import { api } from '../../api/environment-alerting';
 import { alertingSaga } from '../alerting-utils/sagas';
 
-const componentSaga = alertingSaga('ENVIRONMENT_ALERTING', actionCreators, api);
-export default componentSaga;
+const environmentAlertingSaga = alertingSaga(
+  'ENVIRONMENT_ALERTING',
+  actionCreators,
+  api
+);
+export default environmentAlertingSaga;
